+++ card/lib/card.rb
@@ -95,14 +95,14 @@ ActiveSupport.run_load_hooks(:before_card, self)
 # {Card::Auth More on accounts}
 #
 class Card < ApplicationRecord
-  #require_dependency "card/mark"
+  Card::Mark
+  Card::Env
+  ## require_dependency "card/mark"
   extend Mark
 
   Card::Name
   Card::Codename
-  ::Card::Query
-  binding.pry
-  Card::Query::MODIFIERS
+  Card::Query
   Card::Format
   Card::Error
   Card::Auth
@@ -117,23 +117,23 @@ class Card < ApplicationRecord
   Card::Act_manager
   Card::Dirty
   Card::Layout
-  #require_dependency "card/name"
-  # require_dependency "card/codename"
-  # require_dependency "card/query"
-  # require_dependency "card/format"
-  # require_dependency "card/error"
-  # require_dependency "card/auth"
-  # require_dependency "card/mod"
-  # require_dependency "card/content"
-  # require_dependency "card/action"
-  # require_dependency "card/act"
-  # require_dependency "card/change"
-  # require_dependency "card/reference"
-  # require_dependency "card/subcards"
-  # require_dependency "card/view"
-  # require_dependency "card/act_manager"
-  # require_dependency "card/dirty"
-  # require_dependency "card/layout"
+  ## require_dependency "card/name"
+  # # require_dependency "card/codename"
+  # # require_dependency "card/query"
+  # # require_dependency "card/format"
+  # # require_dependency "card/error"
+  # # require_dependency "card/auth"
+  # # require_dependency "card/mod"
+  # # require_dependency "card/content"
+  # # require_dependency "card/action"
+  # # require_dependency "card/act"
+  # # require_dependency "card/change"
+  # # require_dependency "card/reference"
+  # # require_dependency "card/subcards"
+  # # require_dependency "card/view"
+  # # require_dependency "card/act_manager"
+  # # require_dependency "card/dirty"
+  # # require_dependency "card/layout"
 
   has_many :references_in,  class_name: :Reference, foreign_key: :referee_id
   has_many :references_out, class_name: :Reference, foreign_key: :referer_id
