+++ card/lib/card/act_manager.rb
@@ -1,4 +1,4 @@
-require_dependency "card/act_manager/stage_director"
+# require_dependency "card/act_manager/stage_director"
 
 class Card
   # Manages the whole process of creating an {act Card::Act} ie. changing
