+++ card/lib/card/auth.rb
@@ -1,9 +1,9 @@
 # -*- encoding : utf-8 -*-
 
-require_dependency "card/auth/permissions"
-require_dependency "card/auth/proxy"
-require_dependency "card/auth/setup"
-require_dependency "card/auth/token"
+# require_dependency "card/auth/permissions"
+# require_dependency "card/auth/proxy"
+# require_dependency "card/auth/setup"
+# require_dependency "card/auth/token"
 
 class Card
   # Singleton methods for account authentication and contextualization.
