+++ card/lib/card/content/diff/lcs.rb
@@ -1,4 +1,4 @@
-require_dependency "card/content/diff/processor"
+# require_dependency "card/content/diff/processor"
 
 class Card
   class Content
