+++ card/lib/card/name.rb
@@ -1,6 +1,6 @@
 # -*- encoding : utf-8 -*-
 
-require_dependency "card/env"
+# require_dependency "card/env"
 
 require "cardname"
