+++ card/lib/card/query.rb
@@ -34,7 +34,7 @@ class Card
   # Each condition is either a SQL-ready string (boo) or an Array in this form:
   #    [ field_string_or_sym, Card::Value::Query object ]
   module Query
-    # require_dependency "card/query/card_query"
+    # # require_dependency "card/query/card_query"
     Card::Query::CardQuery
     binding.pry
     # After conversion, ATTRIBUTES is a Hash where the key is the attribute
