+++ card/lib/card/set/format.rb
@@ -58,8 +58,8 @@ class Card
     # {AbstractFormat#before before}) are defined in {AbstractFormat}. These methods are
     # always available in format blocks.
     module Format
-      require_dependency "card/set/format/haml_paths"
-      require_dependency "card/set/format/abstract_format"
+      # require_dependency "card/set/format/haml_paths"
+      # require_dependency "card/set/format/abstract_format"
 
       # define format behavior within a set module
       def format *format_names, &block
