+++ card/lib/card/view.rb
@@ -1,7 +1,7 @@
-require_dependency "card/view/cache"
-require_dependency "card/view/options"
-require_dependency "card/view/classy"
-require_dependency "card/view/permission"
+# require_dependency "card/view/cache"
+# require_dependency "card/view/options"
+# require_dependency "card/view/classy"
+# require_dependency "card/view/permission"
 
 class Card
   # Card::View manages {Options view options}, {Cache view caching}, and
