+++ card/mod/admin/set/self/version.rb
@@ -1,6 +1,6 @@
 # -*- encoding : utf-8 -*-
 
-require_dependency "card/version"
+# require_dependency "card/version"
 
 def ok_to_read
   true
