+++ card/mod/core/chunk/link.rb
@@ -1,6 +1,6 @@
 # -*- encoding : utf-8 -*-
 
-require_dependency File.expand_path("../reference", __FILE__)
+# require_dependency File.expand_path("../reference", __FILE__)
 
 module Card::Content::Chunk
   class Link < Reference
