+++ card/mod/core/format/html_format.rb
@@ -1,6 +1,6 @@
 # -*- encoding : utf-8 -*-
 
-# require_dependency "card/content/diff"
+# # require_dependency "card/content/diff"
 
 class Card
   class Format
